refactor: cohérence backend et corrections sécurité
- Sécurité : garde production sur SECRET_KEY et POSTGRES_PASSWORD, credentials seed via env vars, fix token Redis orphelin sur échec SMTP
- Schémas Pydantic : EmailStr sur tous les champs email, min_length=8 sur les mots de passe
- Modèles : Float → Numeric(10,2) sur toutes les colonnes monétaires + migration alembic
- Alembic : import des modèles order et tournament_registration dans env.py
- Routers : messages d'erreur uniformisés en français, status constants dans uploads.py, variable auth ignorée unifiée en _
- main.py : sys.exit(1) si DB injoignable au démarrage, logs sur les erreurs health check
- requirements.txt : suppression de httpx (inutilisé)

+++ backend/app/models/product.py
@@ -1,7 +1,7 @@
 import enum
 from datetime import datetime
 
-from sqlalchemy import DateTime, Enum, Float, Integer, String, Text, func
+from sqlalchemy import DateTime, Enum, Integer, Numeric, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.database import Base
@@ -19,7 +19,7 @@ class Product(Base):
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
     name: Mapped[str] = mapped_column(String(255), nullable=False)
     description: Mapped[str | None] = mapped_column(Text, nullable=True)
-    price: Mapped[float] = mapped_column(Float, nullable=False)
+    price: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
     category: Mapped[ProductCategory] = mapped_column(
         Enum(ProductCategory, name="productcategory"), nullable=False
     )
